requirements.txt
Pillow==9.2.0
requests==2.28.1
streamlit==1.16.0
streamlit-lottie==0.0.3
streamlit-timeline==0.0.2